fixed issues with mean and std, also started working on PCA
